requirements.txt
Flask-WTF==1.0.1
email-validator==1.2.1
requests==2.27.1
python-dotenv==0.20.0
gunicorn==20.1.0
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.1
Bootstrap-Flask==2.1.0
